split the editor toolbar files
