+++ client/src/app/components/App/Canvas/EditorContainer/editorContainer.scss
@@ -1,67 +1,89 @@
 @import 'variables';
 @import 'mixins';
 
-.editor__console{
-  border-radius: 0 0  $medium-border-radius $medium-border-radius;
-  height: 100%;
-  overflow-y: scroll;
-  color:$white;
-  padding: 0 $small-padding;
+.editor
+{
+  &__console{
+    border-radius: 0 0  $medium-border-radius $medium-border-radius;
+    height: 100%;
+    overflow-y: scroll;
+    color:$white;
+    padding: 0 $small-padding;
+  }
+
+  &__container {
+    position: relative;
+    width: 100%;
+    height: 100%;
+    border:1px solid $g5;
+    box-sizing: border-box;
+  }
+
+  &__vertical-container {
+    height: 100%;
+    display: flex;
+  }
+
+  &__sub-container {
+    display: flex;
+    flex-direction: row;
+    border: none;
+  }
+
+  &__input {
+    height: 100%;
+    border-right: $small-border solid $gray;
+    overflow:scroll;
+    &-tabbed {
+      width: 100%;
+    }
+  }
+
+  &__output-overlay {
+    height: 100%;
+    width: 100%;
+    position: absolute;
+    display: none;
+    &--show { // sass-lint:disable-line class-name-format
+      display: block;
+    }
+  }
+
+  &__output {
+    width: 100%;
+    height: 100%;
+    overflow:scroll;
+    background-color: $white;
+    padding: 0;
+    margin: 0;
+    z-index: $z-index-layer-1;
+    &--short {
+      height: calc(100% - #{$console-output-height} - #{$console-heading-height});
+    }
+  }
+
+  &__total-container{
+    height: 100%;
+    display: flex;
+    flex-direction: column;
+    margin: 0;
+    width: auto;
+  }
+
+  &__output-text {
+    color:$white;
+  }
+
+  &__tabbed {
+    .Pane1 {
+      width: 100% !important;
+    }
+  }
 }
 
 .CodeMirror-scroll { // sass-lint:disable-line class-name-format
   z-index: $z-index-layer-0;
 }
-
-.editor__container {
-  position: relative;
-  width: 100%;
-  height: calc(100% - 2 * #{$editor-toolbar-height});
-  border:1px solid $g5;
-  box-sizing: border-box;
-}
-
-.editor__sub-container {
-  display: flex;
-  flex-direction: row;
-  border: none;
-}
-
-.editor__input {
-  height: 100%;
-  border-right: $small-border solid $gray;
-  overflow:scroll;
-  &-tabbed {
-    width: 100%;
-  }
-}
-.editor__output-overlay {
-  height: 100%;
-  width: 100%;
-  position: absolute;
-  display: none;
-  &--show { // sass-lint:disable-line class-name-format
-    display: block;
-  }
-}
-
-.editor__output {
-  width: 100%;
-  height: 100%;
-  overflow:scroll;
-  background-color: $white;
-  padding: 0;
-  margin: 0;
-  z-index: $z-index-layer-1;
-  &--short {
-    height: calc(100% - #{$console-output-height} - #{$console-heading-height});
-  }
-}
-
-.editor__output-text {
-  color:$white;
-}
-
 
 #code-output{
   background-color: $white;
@@ -75,20 +97,6 @@
 canvas{
   padding: 0;
   margin: 0;
-}
-
-.editor__total-container{
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-  margin: 0;
-  width: auto;
-}
-
-.editor__tabbed {
-  .Pane1 {
-    width: 100% !important;
-  }
 }
 
 .editor-toolbar__button-container-left {
