+++ client/src/app/components/App/Shared/EditorComponents/EditorFiles/editorFiles.scss
@@ -1,0 +1,144 @@
+@import 'variables';
+@import 'mixins';
+@import 'base';
+
+.beta-tag {
+  @extend %beta-tag;
+}
+
+.editor__dark {
+  .editor-toolbar {
+    &__container {
+      background-color: $light-gray;
+      background: $dark-gray; /* For browsers that do not support gradients */
+      .editor-toolbar__files {
+        background-color: $g2;
+        .editor-toolbar__file {
+          background-color: $g4;
+          border: $small-border solid $g2;
+          &:hover, &--selected { // sass-lint:disable-line class-name-format
+            background-color: $g2;
+          }
+          &-button {
+            white-space: nowrap;
+            color: $white;
+            box-sizing: border-box;
+            &:hover, &--selected { // sass-lint:disable-line class-name-format
+              color: $white;
+            }
+            &-static {
+              overflow: hidden;
+              color: $g5;
+              &:hover {
+                color: $g5;
+                cursor: default;
+              }
+            }
+          }
+        }
+      }
+    }
+  }
+}
+
+.editor__light {
+  .editor-toolbar {
+    &__container {
+      .editor-toolbar__files {
+        background-color: $g5;
+        .editor-toolbar__file {
+          background-color: $g8;
+          border: $small-border solid $g5;
+          &:hover, &--selected { // sass-lint:disable-line class-name-format
+            background-color: $g4;
+            color: $g8;
+          }
+          &-button {
+            color: $g3;
+            box-sizing: border-box;
+            & #ShapeCNTRL {
+              fill: $g3;
+            }
+            &:hover, &--selected { // sass-lint:disable-line class-name-format
+              color: $g8;
+              & #ShapeCNTRL {
+                fill: $g8;
+              }
+            }
+            &-static {
+              overflow: hidden;
+              color: $g5;
+              &:hover {
+                color: $g5;
+                cursor: default;
+              }
+            }
+          }
+        }
+      }
+    }
+  }
+}
+
+.editor-toolbar {
+  &__image-upload {
+    position: absolute;
+    background: $lighter-gray;
+    z-index: $z-index-layer-4;
+    left: calc(50% - #{$image-notice-width/2});
+    top: 30%;
+    width: $image-notice-width;
+    padding: $large-padding;
+    border-radius: $small-border-radius;
+    box-shadow: 0px 0px $border-shadow $g4;
+  }
+  &__image-close {
+    font-size: $medium-font-size;
+    border: 0;
+    position: absolute;
+    right: $medium-padding;
+    top: $medium-padding;
+    background: transparent;
+    color: $g4;
+    border: none;
+    &:hover {
+      background: transparent;
+      color: $g2;
+      border: none;
+    }
+  }
+  &__image-notice {
+    text-align: center;
+    text-transform: capitalize;
+  }
+
+  &__container {
+    .editor-toolbar__files {
+      overflow-y: scroll;
+      display: flex;
+      width: fit-content;
+      flex-direction: column;
+      margin: 0;
+      padding: 0 !important;
+      .editor-toolbar__file {
+        padding: 0px 3px !important;
+        text-decoration: none;
+        list-style: none;
+        width: auto;
+        display: flex;
+        flex-direction: row;
+        &-button {
+          border-radius: 0;
+          border: none;
+          box-shadow: none;
+          height: 100%;
+          background: transparent;
+          &:hover, &--selected { // sass-lint:disable-line class-name-format
+            border: none;
+            background: transparent;
+          }
+        }
+      }
+    }
+  }
+}
